+++ config/protocols.cfg
@@ -570,7 +570,7 @@
         </protocol>
         <protocol name="sub_ecu_hitachi_m32r_can">
             <ecu>WA12212970WWW</ecu>
-            <mcu>M32R_512KB</mcu>
+            <mcu>M32R_512KB_1block</mcu>
             <mode>OBD2</mode>
             <checksum>yes</checksum>
             <read>no</read>
@@ -588,7 +588,7 @@
             <ecu>M32176F4</ecu>
             <mcu>M32R_512KB</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>yes</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -770,7 +770,7 @@
             <ecu>Denso SH7055</ecu>
             <mcu>SH7055</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -788,7 +788,7 @@
             <ecu>Denso SH7058</ecu>
             <mcu>SH7058</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -806,7 +806,7 @@
             <ecu>Denso SH7055</ecu>
             <mcu>SH7055</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -824,7 +824,7 @@
             <ecu>Denso SH7058</ecu>
             <mcu>SH7058</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -843,7 +843,7 @@
             <ecu>Denso SH7058</ecu>
             <mcu>SH7058</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -855,13 +855,13 @@
             <cal_id_length>5</cal_id_length>
             <kernel>ssmk_can_tp_sh7058.bin</kernel>
             <kernel_addr>0xFFFF3000</kernel_addr>
-            <description>Subaru EEPROM Denso SH7058 CAN-TP</description>
+            <description>Subaru EEPROM Denso SH7058 CAN</description>
         </protocol>
         <protocol name="sub_ecu_eeprom_denso_sh7058_can_diesel">
             <ecu>Denso SH7058</ecu>
             <mcu>SH7058</mcu>
             <mode>OBD2</mode>
-            <checksum>n/a</checksum>
+            <checksum>no</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -873,7 +873,7 @@
             <cal_id_length>5</cal_id_length>
             <kernel>ssmk_can_tp_sh7058d_euro4.bin</kernel>
             <kernel_addr>0xFFFF4000</kernel_addr>
-            <description>Subaru EURO4 Diesel EEPROM Denso SH7058 CAN-TP</description>
+            <description>Subaru EEPROM Denso SH7058 CAN EURO4 Diesel</description>
         </protocol>
 
 
@@ -881,7 +881,7 @@
             <ecu>Unisia Jecs</ecu>
             <mcu>M3779x</mcu>
             <mode>SSM1</mode>
-            <checksum>no</checksum>
+            <checksum>n/a</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -893,13 +893,13 @@
             <cal_id_length>5</cal_id_length>
             <kernel></kernel>
             <kernel_addr>0x00000000</kernel_addr>
-            <description>1992 SSM ECU module</description>
+            <description>Subaru Impreza, Legacy, SVX 1992-1996 SSM</description>
         </protocol>
         <protocol name="sub_ecu_unisia_jecs_m3775x">
             <ecu>Unisia Jecs</ecu>
             <mcu>M3775x</mcu>
             <mode>SSM1</mode>
-            <checksum>no</checksum>
+            <checksum>n/a</checksum>
             <read>yes</read>
             <test_write>no</test_write>
             <write>no</write>
@@ -911,7 +911,7 @@
             <cal_id_length>5</cal_id_length>
             <kernel></kernel>
             <kernel_addr>0x00000000</kernel_addr>
-            <description>1997 OBD2 SSM ECU module</description>
+            <description>Subaru Impreza, Legacy, SVX 1997-1998 OBD2 SSM</description>
         </protocol>
     </protocols>
