+++ modules/biu/biu_operations_subaru.h
@@ -13,10 +13,10 @@
 #include <QDialog>
 
 #include <serial_port/serial_port_actions.h>
-#include <biu_ops_subaru_switches.h>
-#include <biu_ops_subaru_data.h>
-#include <biu_ops_subaru_input1.h>
-#include <biu_ops_subaru_input2.h>
+#include <modules/biu/biu_ops_subaru_switches.h>
+#include <modules/biu/biu_ops_subaru_data.h>
+#include <modules/biu/biu_ops_subaru_input1.h>
+#include <modules/biu/biu_ops_subaru_input2.h>
 
 QT_BEGIN_NAMESPACE
 namespace Ui
